pyproject.toml
[project]
name = "vibe-coding-community"
version = "0.1.0"
description = "Vibe Coding互助社区 - 连接AI编程求助者与解决者的平台"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.4",
    "jinja2-fragments>=1.5.2",
    "python-multipart>=0.0.18",
    "supabase>=2.11.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]
